# compiler(amd64): optimize mem to register operations