@ncruces 2023-06-15
If you think this takes away precious attention span, this one is probably not worth it.