@mathetake 2023-06-15
I would suggest that this would be the last change to the existing code generation code until the optimizing compiler is up for collaboration. wdyt? @evacchi @codefromthecrypt @ncruces @achille-roussel