@mathetake 2023-06-15
> If you think this takes away precious attention span, this one is probably not worth it.

the problem here is that review might be fine, but I cannot be sure if it wouldn't have a bug because we are not running large fuzzing on a daily basis.