@mathetake 2023-06-15
e.g. https://github.com/tetratelabs/wazero/tree/main/internal/integration_test/fuzzcases/testdata these are the bugs we've found only after multiple days of fuzzing