@mathetake 2023-06-15
I guess it's time to setup the long fuzzing CI for both arm64/amd64 which we run every week or so plus before the release.