@ncruces 2023-06-15
I actually don't mind outright closing this one.
I was working on it for about a day, thought it might help, but it barely ticks on larger benchmarks.

The other ones I consider "instructive":
- optimize for the happy path, those gave big gains;
- eliminating bounds checks will offer the most gains;
- all other changes where to macro ops: code them in C, go to [compiler explorer](https://godbolt.org/), generate same code.

I imagine those changes are useful (1) as a stopgap (2) because they inform the new backend.
This one is completely obviated by an SSA compiler that does more informed register allocation.
It has zero value for the future.