@mathetake 2023-06-15
yeah agreed, these changes could be helpful in the new backend as well, and also agree that this one is destined to be obviated by the new one since this optimization depends on the current compiler's register usage. Thank you for your efforts and dedication!

In any way, for any changes to the existing backend, I definitely would like to do a comprehensive fuzzing before each release due to the reason above.