@mathetake 2023-06-16
let me clarify the state in case I caused some confusion here: @evacchi will setup the heavy fuzzing pipeline. After that we will continue any patches like this to the existing backed, which of course will be used by the new compiler as well. That way, I can confidently say LGTM to your awesome patches @ncruces. Thanks again for your dedication and time! 🏅